Cookie-based token storage (without DB)
In order to avoid XSS and CSRF attack, we need to store
a unique session ID in a http-only cookie.
For more details, see
http://www.redotheweb.com/2015/11/09/api-security.html

The only difference with the specified method is that
we do not have a DB table to store the unique session ID.
Instead, we encrypt the JWT token with a new secret key
and store it in the cookie.
This way, at each API call, we can crypt again the specified JWT
and check wether it corresponds to the cookie token.
Since the JWT is deleted at end of the session, so is the cookie token.